app/requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
uvicorn==0.27.1
httpx==0.27.0
pydub==0.25.1
aiohttp==3.9.3
python-multipart==0.0.9
